refactor: count 메서드 관련 반환 타입을 Integer에서 Long으로 변경

+++ backend/src/main/java/moaon/backend/article/dto/ArticleSectorCount.java
@@ -4,10 +4,10 @@ import moaon.backend.article.domain.Sector;
 
 public record ArticleSectorCount(
         String sector,
-        int count
+        long count
 ) {
 
-    public static ArticleSectorCount of(Sector sector, int count) {
+    public static ArticleSectorCount of(Sector sector, long count) {
         return new ArticleSectorCount(sector.getName(), count);
     }
 }
